fix Ledger wallet selection
As part of 2fbe81a78a4deced364ad461968ecb3c26d53be3 the wallet providers
were updated to a new source. This had the side effect of changing the
wallet that is selected when connecting using the 'Ledger' provider.
Before the update, the wallet at derivation path 44'/501' was selected,
also known as the root Solana account. After the update, the derivation
path default got changed to 44'/501'/0'/0'.
This patch restores the original derivation path 44'/501' for now to
avoid breaking wallet functionality for users until the derivation path
can be configured through the UI.